Sacando las vistas y las tasks a una nueva app
