+++ elections/urls.py
@@ -4,15 +4,13 @@ from haystack.views import SearchView
 from elections.forms import ElectionForm
 from elections.views import ElectionsSearchByTagView, HomeView, ElectionDetailView,\
     CandidateDetailView, SoulMateDetailView, FaceToFaceView, AreaDetailView, \
-    CandidateFlatPageDetailView, ElectionRankingView, QuestionsPerCandidateView
+    CandidateFlatPageDetailView, ElectionRankingView
 from sitemaps import *
 
 from django.views.decorators.cache import cache_page
-from elections.preguntales_views import MessageDetailView, ElectionAskCreateView, AnswerWebHook
 
 media_root = getattr(settings, 'MEDIA_ROOT', '/')
 
-new_answer_endpoint = r"^new_answer/%s/?$" % (settings.NEW_ANSWER_ENDPOINT)
 
 sitemaps = {
     'elections': ElectionsSitemap,
@@ -20,7 +18,6 @@ sitemaps = {
 }
 
 urlpatterns = patterns('',
-    url(new_answer_endpoint,AnswerWebHook.as_view(), name='new_answer_endpoint' ),
     url(r'^/?$', cache_page(60 * settings.CACHE_MINUTES)(HomeView.as_view(template_name='elections/home.html')), name='home'),
     url(r'^buscar/?$', SearchView(template='search.html',
             form_class=ElectionForm), name='search'),
@@ -47,22 +44,10 @@ urlpatterns = patterns('',
     url(r'^election/(?P<slug>[-\w]+)/soul-mate/?$',
         SoulMateDetailView.as_view(template_name='elections/soulmate_candidate.html'),
         name='soul_mate_detail_view'),
-    # Preguntales
-    url(r'^election/(?P<election_slug>[-\w]+)/messages/(?P<pk>\d+)/?$',
-        MessageDetailView.as_view(template_name='elections/message_detail.html'),
-        name='message_detail'),
     #ranking
     url(r'^election/(?P<slug>[-\w]+)/ranking/?$',
         cache_page(60 * settings.CACHE_MINUTES)(ElectionRankingView.as_view(template_name='elections/ranking_candidates.html')),
         name='ranking_view'),
-    url(r'^election/(?P<election_slug>[-\w]+)/(?P<slug>[-\w]+)/questions?$',
-        QuestionsPerCandidateView.as_view(template_name='elections/questions_per_candidate.html'),
-        name='questions_per_candidate'
-        ),
-    #ask
-    url(r'^election/(?P<slug>[-\w]+)/ask/?$',
-        ElectionAskCreateView.as_view(template_name='elections/ask_candidate.html'),
-        name='ask_detail_view'),
 
     url(r'^election/(?P<election_slug>[-\w]+)/(?P<slug>[-\w]+)/?$',
         cache_page(60 * settings.CACHE_MINUTES)(CandidateDetailView.as_view(template_name='elections/candidate_detail.html')),
