+++ votainteligente/urls.py
@@ -18,6 +18,7 @@ urlpatterns = patterns('',
     # Uncomment the next line to enable the admin:
     url(r'^admin/', include(admin.site.urls)),
     url(r'^', include('elections.urls')),
+    (r'^preguntales', include('preguntales.urls')),
     ('^pages/', include('django.contrib.flatpages.urls')),
     (r'^tinymce/', include('tinymce.urls')),
     (r'^newsletter/', include('newsletter.urls')),
